requirements.txt
asgiref==3.2.10
Babel==2.8.0
dj-database-url==0.5.0
Django==3.0.7
django-crispy-forms==1.9.1
django-phonenumber-field==4.0.0
django-weekday-field==1.1.0
docutils==0.16
gunicorn==20.0.4
phonenumbers==8.12.6
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.1.0
